// frontend/src/app/shared/parcel-status.enum.ts
export enum ParcelStatus {
    PENDING = 'PENDING',
    IN_TRANSIT = 'IN_TRANSIT',
    READY_FOR_COLLECTION = 'READY_FOR_COLLECTION',
    COLLECTED = 'COLLECTED',
    DELIVERED = 'DELIVERED',
    DAMAGED = 'DAMAGED',
    COMPLAINT_BOX = 'COMPLAINT_BOX',
    DELETED = 'DELETED',
    CANCELLED = 'CANCELLED',
}